+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## 0.12.1
-- Added support for additional lock, siren, and sensor attributes/features
+- Added support for additional lock, garage door, siren, and sensor attributes/features
 
 ## 0.12.0
 - Wink fan support